Inline panel commands to make the panel even further self-contained
